chore: disable log
